INI-698 Feat: Add morph relationship for offers, fix formatting inconsistencies, typecast trigger data, and introduce maintenance action for Aurora category offers.

+++ app/Actions/Discounts/Offer/StoreProductCategoryDiscount.php
@@ -70,7 +70,7 @@ class StoreProductCategoryDiscount extends OrgAction
             $modelData,
             'trigger_data',
             [
-                'item_quantity' => Arr::pull($modelData, 'trigger_data_item_quantity')
+                'item_quantity' => (int) Arr::pull($modelData, 'trigger_data_item_quantity')
             ]
         );
